requirements.txt
python-dotenv==1.0.0
telethon==1.34.0
pymongo==4.6.1
flask==3.0.2
python-telegram-bot==20.8
cryptg==0.4.0  # For faster downloads/uploads
ujson==5.9.0  # For better JSON performance
dnspython==2.6.1  # Required by pymongo
aiohttp==3.9.3  # Required by telethon
